Split data table actions into sub modules
