+++ webapp/app/assessment/components/dataTable/actions/index.js
@@ -1,0 +1,11 @@
+export {
+  assessmentSectionDataUpdate,
+  postTableData,
+  updateTableData,
+  updateTableDataCell,
+  updateTableWithOdpCell,
+} from './update'
+
+export { assessmentSectionDataGeneratingValuesUpdate, generateTableData } from './generate'
+
+export { fetchTableData } from './read'
